Actor class reorganization
Move actors and sheet classes to their own folder to clean up root ts
folder.
