+++ module/actors/sheets/character-sheet.ts
@@ -1,22 +1,22 @@
-import { BWActor, NewItemData } from "./bwactor.js";
+import { BWActor, NewItemData } from "../bwactor.js";
 import { BWActorSheet } from "./bwactor-sheet.js";
-import * as constants from "./constants.js";
+import * as constants from "../../constants.js";
 import {
     BWItemData,
-} from "./items/item.js";
-import { handleRollable } from "./rolls/rolls.js";
-import { CharacterBurnerDialog } from "./dialogs/characterBurner.js";
-import { addNewItem } from "./dialogs/importItemDialog.js";
-import { BWCharacter } from "./character.js";
-import { byName } from "./helpers.js";
-import { ArmorRootData } from "./items/armor.js";
-import { MeleeWeaponRootData, MeleeWeaponData } from "./items/meleeWeapon.js";
-import { RangedWeaponRootData } from "./items/rangedWeapon.js";
-import { RelationshipData } from "./items/relationship.js";
-import { ReputationDataRoot } from "./items/reputation.js";
-import { SkillDataRoot, Skill } from "./items/skill.js";
-import { SpellDataRoot, Spell } from "./items/spell.js";
-import { TraitDataRoot, Trait } from "./items/trait.js";
+} from "../../items/item.js";
+import { handleRollable } from "../../rolls/rolls.js";
+import { CharacterBurnerDialog } from "../../dialogs/characterBurner.js";
+import { addNewItem } from "../../dialogs/importItemDialog.js";
+import { BWCharacter } from "../../character.js";
+import { byName } from "../../helpers.js";
+import { ArmorRootData } from "../../items/armor.js";
+import { MeleeWeaponRootData, MeleeWeaponData } from "../../items/meleeWeapon.js";
+import { RangedWeaponRootData } from "../../items/rangedWeapon.js";
+import { RelationshipData } from "../../items/relationship.js";
+import { ReputationDataRoot } from "../../items/reputation.js";
+import { SkillDataRoot, Skill } from "../../items/skill.js";
+import { SpellDataRoot, Spell } from "../../items/spell.js";
+import { TraitDataRoot, Trait } from "../../items/trait.js";
 
 export class BWCharacterSheet extends BWActorSheet {
     get actor(): BWCharacter {
