+++ module/helpers.ts
@@ -1,4 +1,4 @@
-import { TracksTests } from "./bwactor.js";
+import { TracksTests } from "./actors/bwactor.js";
 import { ArmorRootData } from "./items/armor.js";
 import { BWItem, DisplayClass, ItemType } from "./items/item.js";
 import { Skill } from "./items/skill.js";
