+++ module/items/trait.ts
@@ -1,4 +1,4 @@
-import { RollModifier } from "../bwactor.js";
+import { RollModifier } from "../actors/bwactor.js";
 import { ItemType, HasPointCost, BWItemData, BWItem } from "./item.js";
 
 export class Trait extends BWItem {
