+++ module/rolls/rerollTraitors.ts
@@ -1,5 +1,5 @@
 import { TestString } from "../helpers.js";
-import { Ability, BWActor, TracksTests } from "../bwactor.js";
+import { Ability, BWActor, TracksTests } from "../actors/bwactor.js";
 import * as helpers from "../helpers.js";
 import { getNoDiceErrorDialog, RerollMessageData, rollDice, templates } from "./rolls.js";
 import { BWCharacter } from "../character.js";
