+++ module/rolls/rolls.ts
@@ -1,4 +1,4 @@
-import { BWActor, RollModifier, TracksTests } from "../bwactor.js";
+import { BWActor, RollModifier, TracksTests } from "../actors/bwactor.js";
 import * as helpers from "../helpers.js";
 import { handleAttrRollEvent } from "./rollAttribute.js";
 import { handleCirclesRollEvent } from "./rollCircles.js";
@@ -11,8 +11,8 @@ import { handleArmorRollEvent } from "./rollArmor.js";
 import { handleWeaponRollEvent } from "./rollWeapon.js";
 import { handleSpellRollEvent } from "./rollSpell.js";
 import { handleSpellTaxRoll } from "./rollSpellTax.js";
-import { BWCharacterSheet } from "../character-sheet.js";
-import { NpcSheet } from "../npc-sheet.js";
+import { BWCharacterSheet } from "../actors/sheets/character-sheet.js";
+import { NpcSheet } from "../actors/sheets/npc-sheet.js";
 import { Possession } from "../items/possession.js";
 
 export async function handleRollable(
